feat(migrations): add user_collections table to the database schema to store user collections
This commit adds a new migration file that creates a new table called user_collections in the database schema. This table has columns for id, userId, typeId, title, description, visibility, and timestamps. The userId and typeId columns are foreign keys that reference the id column of the users and collections_types tables respectively. The visibility column is an enum that specifies the visibility of the collection. This table will be used to store user collections.