feat(migrations): add logs table and triggers to log changes in user-related tables
This migration adds a new table called logs to store changes in user-related tables. It also creates a function called custom_log that inserts a new row in the logs table every time a new row is inserted in one of the user-related tables. Finally, it creates a trigger for each user-related table that calls the custom_log function after a new row is inserted. This way, we can keep track of changes made to user-related data.